File transfer works, tested end to end over the air
I got up to 6560 bps on a 3684 byte file with 194 byte chunk size.

I have not calculated the theoretical maximum speed with N sized
packets. It's a bit complicated, because it involves hold times before
and after the packet, in addition to the AX.25 header and bit
stuffing.

Larger packets reduces overhead, but the Kenwood TH-D74 crashes if the
packets are too big. I think it may also drop and/or corrupt packets
if I send them too quickly.

But this test is at least 68% of the way to 9600bps. And it's 5.5x
faster than the theoretical 1200bps of sound card based
implementations.